+++ main.py
@@ -1,10 +1,9 @@
 import os
+import time
 import traceback
 from Queue import Queue
 from multiprocessing import Value
 from threading import Thread
-
-import time
 
 import bot_logger
 import crypto