+++ user_function.py
@@ -1,6 +1,5 @@
-import datetime
 import json
-import random
+
 from tinydb import TinyDB, Query
 
 import bot_logger
@@ -29,12 +28,15 @@ def add_user(user, address):
 
 def get_user_address(user):
     user_list = get_users()
-    return user_list[user]
+    try:
+        return user_list[user]
+    except KeyError:
+        return None
 
 
 def user_exist(user):
     user_list = get_users()
-    if user in user_list.keys():
+    if user.lower() in map(unicode.lower, user_list.keys()):
         return True
     else:
         return False