+++ CHANGELOG.md
@@ -1,10 +1,20 @@
-# v1.1 - xx/06/2017
+# v1.2 - x/06/2017
+- Add error on invalid currency
+- Fix an error on tip to lowercase user registered
+- Fix an error on pending tips
+- Fix an error on withdraw in history
+
+# v1.1 - 16/06/2017
 - Add support of new syntax for tip command #57
 - Add donate command
 - Check wallet pass phrase on startup
 - Fix bug on update status of tip in history
 - Remove extra decimal on date in history
 - Update log rotation file
+- Simplify +info and +balance commands
+- Fix an issue when register in comments
+- Add limit on number of entry returned by +history
+- Add link to chain.so with TxID in +history
 
 # v1.0 - 10/06/2017
 - First release